# v2/api/contract/control-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Norn Control Protocol
  version: 1.8.0
  description: Versioned contract for native and automated Norn control clients.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/github-actions/exchange:
    post:
      operationId: exchangeGitHubActionsOIDC
      description: Exchanges a verified GitHub Actions OIDC bearer assertion for a short-lived Norn release token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [scope, environment]
              properties:
                scope: { type: string }
                app: { type: string, description: 'Required for release scopes; omitted for fleet:operate.' }
                environment: { type: string }
                intent: { type: string }
              allOf:
                - if: { properties: { scope: { enum: ['release:stage', 'release:qualify', 'release:promote', 'release:rollback'] } }, required: [scope] }
                  then: { required: [app] }
      responses:
        '201': { description: Exchange token created }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    Problem:
      description: Stable machine-readable error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { $ref: '#/components/schemas/ErrorCode' }
        detail: { type: string }
        requestId: { type: string }
        error: { type: string, deprecated: true }
        eventBounds: { $ref: '#/components/schemas/EventBounds' }
    ErrorCode:
      type: string
      description: Stable machine-readable code. Clients must preserve unknown future values.
      x-extensible-enum:
        - unauthorized
        - authenticated_principal_required
        - insufficient_scope
        - invalid_request
        - invalid_app_id
        - app_not_found
        - app_discovery_failed
        - invalid_deployment_id
        - deployment_not_found
        - deployment_store_unavailable
        - deployment_history_read_failed
        - deployment_steps_read_failed
        - service_manifest_read_failed
        - release_list_failed
        - invalid_event_cursor
        - invalid_event_filter
        - invalid_heartbeat
        - event_store_unavailable
        - event_cursor_gap
        - event_cursor_ahead
        - operation_not_found
        - operation_lookup_failed
        - operation_create_failed
        - operation_cancel_failed
        - operation_not_cancelable
        - invalid_idempotency_key
        - idempotency_key_reused
        - app_operation_active
        - postgres_not_configured
        - migration_not_configured
        - confirmation_required
        - invalid_snapshot_retention
        - invalid_snapshot_restore
        - invalid_snapshot_timestamp
        - invalid_snapshot_identifier
        - invalid_migration_request
        - invalid_migration_ref
        - invalid_rollback_request
        - invalid_rollback_region
        - rollback_current_deployment_missing
        - rollback_target_missing
        - operation_already_finished
        - invalid_platform_ref
        - invalid_upgrade_mode
        - invalid_drain_mode
        - invalid_release_sha
        - invalid_idempotency_key
        - invalid_device_name
        - invalid_scope
        - invalid_token_note
        - token_expired
        - token_registry_unavailable
        - scope_not_requested
        - invalid_public_key
        - enrollment_failed
        - enrollment_store_unavailable
        - enrollment_list_failed
        - enrollment_not_found
        - enrollment_lookup_failed
        - enrollment_not_pending
        - enrollment_not_approved
        - invalid_enrollment_verifier
        - enrollment_already_exchanged
        - enrollment_rate_limited
        - device_not_found
        - device_not_active
        - device_lookup_failed
        - device_list_failed
        - device_revoke_failed
        - managed_token_required
        - token_not_managed
        - token_issue_failed
        - token_revoke_failed
        - managed_device_required
        - invalid_step_up_request
        - invalid_step_up_proof
        - step_up_key_required
        - step_up_failed
        - step_up_lookup_failed
        - step_up_challenge_not_found
        - step_up_challenge_inactive
        - step_up_rate_limited
        - nomad_unavailable
        - exec_step_up_required
        - step_up_already_consumed
        - invalid_exec_request
        - invalid_exec_argv
        - invalid_terminal_size
        - running_allocation_not_found
        - exec_session_create_failed
        - exec_session_not_found
        - exec_session_lookup_failed
        - exec_session_list_failed
        - exec_session_forbidden
        - exec_session_inactive
        - exec_session_not_cancelable
        - exec_session_cancel_failed
        - invalid_exec_encoding
        - exec_session_limit_reached
        - exec_session_rate_limited
        - mutation_audit_unavailable
        - mutation_audit_unsigned
        - mutation_audit_read_failed
        - production_substrate_unready
        - recovery_drills_unavailable
        - recovery_drills_read_failed
        - recovery_drill_create_failed
        - recovery_drill_not_running
        - recovery_drill_complete_failed
        - invalid_recovery_drill_kind
        - invalid_recovery_drill_target
        - invalid_recovery_drill_status
        - invalid_recovery_drill_evidence
        - invalid_limit
        - invalid_validation_request
        - validation_document_required
        - validation_document_too_large
        - invalid_fleet_plan_request
        - fleet_not_configured
        - fleet_config_invalid
        - fleet_config_read_failed
        - fleet_node_pool_not_found
        - fleet_plan_unsafe
        - fleet_plan_store_unavailable
        - fleet_plan_store_failed
        - fleet_plan_finish_failed
        - fleet_plan_read_failed
        - invalid_fleet_plan_id
        - fleet_plan_not_found
        - invalid_fleet_reconciliation
        - fleet_reconciliation_out_of_order
        - fleet_reconciliation_store_unavailable
        - fleet_reconciliation_store_failed
        - fleet_reconciliation_read_failed
        - invalid_fleet_runner_attempt
        - invalid_fleet_runner_attempt_id
        - fleet_runner_attempt_not_found
        - fleet_runner_attempt_exists
        - fleet_runner_retry_required
        - fleet_runner_plan_already_complete
        - fleet_runner_attempt_store_unavailable
        - fleet_runner_attempt_store_failed
        - fleet_runner_attempt_read_failed
        - fleet_runner_attempt_update_failed
        - fleet_runner_attempt_binding_conflict
        - fleet_runner_attempt_revision_conflict
        - invalid_fleet_runner_heartbeat
        - fleet_runner_heartbeat_out_of_order
        - invalid_fleet_runner_advance
        - fleet_runner_phase_conflict
        - fleet_runner_advance_unproven
        - invalid_fleet_runner_retry
        - fleet_runner_retry_not_allowed
        - invalid_fleet_runner_cancel
        - fleet_runner_cancel_not_allowed
        - fleet_github_not_configured
        - fleet_github_plan_stale
        - fleet_github_pull_request_failed
        - fleet_github_plan_not_ready
        - fleet_github_destructive_ack_required
        - fleet_github_dispatch_failed
        - fleet_github_receipt_failed
        - fleet_plan_invalid
        - invalid_fleet_github_dispatch
        - idempotency_key_reused
...
    EventBounds:
      type: object
      required: [oldestCursor, latestCursor, retainedEvents]
      properties:
        oldestCursor: { type: integer, format: int64 }
        latestCursor: { type: integer, format: int64 }
        retainedEvents: { type: integer, format: int64 }
        oldestTimestamp: { type: string, format: date-time }
        latestTimestamp: { type: string, format: date-time }
...